% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_dicca.R
\name{dicca_predict_scores}
\alias{dicca_predict_scores}
\title{One-step-ahead prediction of DLVs (diagonal G(B))}
\usage{
dicca_predict_scores(model, scores)
}
\arguments{
\item{model}{a fitted DiCCA model object with compact inner models}

\item{scores}{numeric matrix of observed scores to predict from}
}
\value{
matrix of predicted scores
}
\description{
One-step-ahead prediction of DLVs (diagonal G(B))
}
