% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_dicca.R
\name{dicca_scores}
\alias{dicca_scores}
\title{Transform new data to Di(L)V scores using a fitted compact DiCCA model}
\usage{
dicca_scores(model, Xnew)
}
\arguments{
\item{model}{a fitted DiCCA model object}

\item{Xnew}{numeric matrix of new data to transform}
}
\value{
matrix of scores
}
\description{
Transform new data to Di(L)V scores using a fitted compact DiCCA model
}
