% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipls.R
\name{dipls_scores}
\alias{dipls_scores}
\title{Project X to DiPLS scores}
\usage{
dipls_scores(model, Xnew)
}
\arguments{
\item{model}{fitted DiPLS object from \code{\link{dipls}}.}

\item{Xnew}{numeric matrix with the same number of columns as the training
\code{X}.}
}
\value{
Matrix of latent scores (rows correspond to observations).
}
\description{
Project X to DiPLS scores
}
