% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipls.R
\name{predict.dipls}
\alias{predict.dipls}
\title{Predict Y using a fitted DiPLS model}
\usage{
\method{predict}{dipls}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{"dipls"} object.}

\item{newdata}{matrix \code{X} or a list containing an element \code{X}.}

\item{...}{unused.}
}
\value{
Matrix of one-step-ahead predictions for \code{Y}.
}
\description{
Predict Y using a fitted DiPLS model
}
